FutbolVagos/requirements.txt
asgiref==3.8.1
Django==5.1.7
djangorestframework==3.16.0
drf-yasg==1.21.10
inflection==0.5.1
packaging==24.2
psycopg2==2.9.10
pytz==2025.2
PyYAML==6.0.2
sqlparse==0.5.3
tzdata==2025.2
uritemplate==4.1.1
djangorestframework_simplejwt==5.5.0
cryptography==44.0.2
python-jose==3.4.0
python-keycloak==5.5.0
social-auth-app-django==5.4.3